python/destinations/TDengine/requirements.txt
# TODO: remove "@ git+" version of install before merging PR
quixstreams @ git+https://github.com/quixio/quix-streams.git
python-dotenv